+++ game_patch/misc/alpine_settings.h
@@ -1,0 +1,3 @@
+#pragma once
+
+extern bool g_loaded_alpine_settings_file;